requirements.txt
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1732732445543/work
annotated-types @ file:///croot/annotated-types_1709542908624/work
anyio @ file:///croot/anyio_1729121277521/work
archspec @ file:///croot/archspec_1709217642129/work
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///croot/argon2-cffi-bindings_1736182440035/work
asgiref @ file:///croot/asgiref_1724072470684/work
asttokens @ file:///croot/asttokens_1743630435401/work
async-lru @ file:///work/perseverance-python-buildout/croot/async-lru_1701732681408/work
attrs @ file:///croot/attrs_1734533101012/work
babel @ file:///croot/babel_1737454360933/work
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1718029820055/work
bleach @ file:///croot/bleach_1732290411627/work
boltons @ file:///croot/boltons_1737061692168/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1738623731865/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
comm @ file:///croot/comm_1709322850197/work
conda @ file:///croot/conda_1743715288469/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1736524596117/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1739298990519/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1737733694612/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
cryptography @ file:///croot/cryptography_1732129644106/work
debugpy @ file:///croot/debugpy_1736267418885/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
distro @ file:///croot/distro_1714488253808/work
Django @ file:///croot/django_1731615954020/work
djangorestframework==3.16.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema @ file:///croot/python-fastjsonschema_1731939362158/work
frozendict @ file:///croot/frozendict_1713194832637/work
h11 @ file:///croot/h11_1706652277403/work
httpcore @ file:///croot/httpcore_1706728464539/work
httpx @ file:///croot/httpx_1723474802858/work
idna @ file:///croot/idna_1714398848350/work
ipykernel @ file:///croot/ipykernel_1737660677549/work
ipython @ file:///croot/ipython_1744912971218/work
ipython_pygments_lexers @ file:///croot/ipython_pygments_lexers_1744753235686/work
jedi @ file:///croot/jedi_1733987392413/work
Jinja2 @ file:///croot/jinja2_1741710844255/work
json5 @ file:///croot/json5_1730786798687/work
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema @ file:///croot/jsonschema_1728486696720/work
jsonschema-specifications @ file:///work/perseverance-python-buildout/croot/jsonschema-specifications_1701731664072/work
jupyter-events @ file:///croot/jupyter_events_1741184577592/work
jupyter-lsp @ file:///work/perseverance-python-buildout/croot/jupyter-lsp-meta_1707343107456/work
jupyter_client @ file:///croot/jupyter_client_1737570961872/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
jupyter_server @ file:///croot/jupyter_server_1741200034823/work
jupyter_server_terminals @ file:///croot/jupyter_server_terminals_1744706698694/work
jupyterlab @ file:///croot/jupyterlab_1737555423487/work
jupyterlab_pygments @ file:///croot/jupyterlab_pygments_1741124142640/work
jupyterlab_server @ file:///croot/jupyterlab_server_1725865349919/work
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
Markdown==3.8
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1698846045803/work
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib-inline @ file:///work/perseverance-python-buildout/croot/matplotlib-inline_1698864771271/work
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1698845653285/work
menuinst @ file:///croot/menuinst_1738943416351/work
mistune @ file:///croot/mistune_1741124011532/work
nbclient @ file:///croot/nbclient_1741123995822/work
nbconvert @ file:///croot/nbconvert-meta_1741184653315/work
nbformat @ file:///croot/nbformat_1728049424075/work
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
notebook @ file:///croot/notebook_1738159946465/work
notebook_shim @ file:///croot/notebook-shim_1741707758683/work
overrides @ file:///work/perseverance-python-buildout/croot/overrides_1701732220415/work
packaging @ file:///croot/packaging_1734472117206/work
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///croot/parso_1733963305961/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///croot/pluggy_1733169602837/work
prometheus_client @ file:///croot/prometheus_client_1744271615306/work
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
psutil @ file:///croot/psutil_1736367091698/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///croot/pydantic_1734736067156/work
pydantic_core @ file:///croot/pydantic-core_1734726052986/work
Pygments @ file:///work/perseverance-python-buildout/croot/pygments_1698846270603/work
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-json-logger @ file:///croot/python-json-logger_1734370021104/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq @ file:///croot/pyzmq_1734687138743/work
referencing @ file:///work/perseverance-python-buildout/croot/referencing_1701731622327/work
requests @ file:///croot/requests_1730999120400/work
rfc3339-validator @ file:///work/perseverance-python-buildout/croot/rfc3339-validator_1698873973408/work
rfc3986-validator @ file:///work/perseverance-python-buildout/croot/rfc3986-validator_1698874010100/work
rich @ file:///croot/rich_1732638981168/work
rpds-py @ file:///croot/rpds-py_1736541261634/work
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
Send2Trash @ file:///croot/send2trash_1736540790175/work
setuptools==75.8.0
six @ file:///croot/six_1744271502820/work
sniffio @ file:///croot/sniffio_1705431295498/work
soupsieve @ file:///work/perseverance-python-buildout/croot/soupsieve_1698866207280/work
sqlparse @ file:///croot/sqlparse_1734704446487/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
terminado @ file:///work/perseverance-python-buildout/croot/terminado_1698874934826/work
tinycss2 @ file:///croot/tinycss2_1738337643607/work
tornado @ file:///croot/tornado_1733960490606/work
tqdm @ file:///croot/tqdm_1738943501192/work
traitlets @ file:///croot/traitlets_1718227057033/work
truststore @ file:///croot/truststore_1736550121485/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
urllib3 @ file:///croot/urllib3_1737133630106/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings @ file:///work/perseverance-python-buildout/croot/webencodings_1698866454420/work
websocket-client @ file:///croot/websocket-client_1715878298792/work
wheel==0.45.1
zstandard @ file:///croot/zstandard_1731356346222/work
